[#11] refactor: service interface 삭제
